Type-system hardening: enforce mypy strict + reduce type:ignore by 56%
Three coordinated changes restore mypy strict overrides on
`hyplan.aircraft.wind_path` and `hyplan.planning.isochrone` and clean
up the cascade of suppressions they reveal.  All purely
documentation/typing — zero behavioral changes.

mypy strict cleanup (462 errors → 0)
* Re-enabled the strict-mode overrides dropped in v1.6.0.
* Cleaned 462 cascading errors across 56 files in 4 parallel
  agent passes:
  - 169 `[no-untyped-def]` — added function/method signatures.
  - 148 `[type-arg]` — concrete generic args.
  - 120 `[no-untyped-call]` — resolved transitively.
  - 18 `[attr-defined]` + 7 misc.
* Tactical decisions: `assert is not None` narrowing instead of
  Optional weakening; lazy-import helpers annotated `-> Any`;
  `_GriddedWindField` import switched from private re-export to
  canonical module path.

type:ignore documentation
* All 149 pre-existing bare `# type: ignore[code]` comments now
  carry an inline `# reason` explaining why the suppression
  exists.  Format: `# type: ignore[code]  # short reason`.

Structural fixes that eliminated 43 ignores
* `is_waypoint()` typed as `TypeGuard[Waypoint]` → mypy now
  narrows through `if is_waypoint(x):` blocks (17 ignores cleared
  in `planning/engine.py`).
* `_GriddedWindField` slab attributes given explicit class-level
  `npt.NDArray[...]` annotations instead of the
  `self._x = None` initialization pattern (16 + 10 = 26 ignores
  cleared across `winds/gridded.py` and `winds/providers/gfs.py`).

no-any-return reduction (45 → 16)
* Replaced 29 documented suppressions with explicit Python
  coercion at the boundary:
  - `q.m_as(unit)` → `float(q.m_as(unit))` for scalar Quantity.
  - numpy reductions → `float(...)` (atmosphere, sun, geometry,
    lvis, winds/utils, frame_camera).
  - numpy bool comparisons → `bool(...)` (lvis).
  - pandas indexing → `str(...)` / `list(...)` (airports).
  - `ndarray.shape` → `(int(s[0]), int(s[1]))` tuple comprehension.
* The 16 remaining ignores are all genuine library-boundary
  cases (earthengine / earthaccess / rasterio no stubs;
  requests.json() Any returns; np.median ndarray returns; geomag).

Total documented `# type: ignore` count: 149 → 66 (-56%).
All 1697 tests pass; mypy strict-clean; ruff clean.

+++ hyplan/atmosphere.py
@@ -44,10 +44,12 @@ def _pressure_pa(altitude_m: float) -> float:
     """ISA pressure in Pascals at geometric altitude in metres."""
     if altitude_m <= _TROPOPAUSE_M:
         t = _temperature_k(altitude_m)
-        return _P0 * (t / _T0) ** (-_G / (_L * _R))  # type: ignore[no-any-return]
+        return float(_P0 * (t / _T0) ** (-_G / (_L * _R)))
     # Isothermal layer above tropopause
-    return _P_TROPOPAUSE * np.exp(  # type: ignore[no-any-return]
-        -_G / (_R * _T_TROPOPAUSE) * (altitude_m - _TROPOPAUSE_M)
+    return float(
+        _P_TROPOPAUSE * np.exp(
+            -_G / (_R * _T_TROPOPAUSE) * (altitude_m - _TROPOPAUSE_M)
+        )
     )
 
 
@@ -58,7 +60,7 @@ def _density_kgm3(altitude_m: float) -> float:
 
 def _speed_of_sound_ms(altitude_m: float) -> float:
     """Speed of sound in m/s at geometric altitude in metres."""
-    return np.sqrt(_GAMMA * _R * _temperature_k(altitude_m))  # type: ignore[no-any-return]
+    return float(np.sqrt(_GAMMA * _R * _temperature_k(altitude_m)))
 
 
 # ---------------------------------------------------------------------------
@@ -165,4 +167,4 @@ def tas_to_mach(tas: Quantity, altitude: Quantity) -> float:
     tas_ms = tas.m_as(ureg.meter / ureg.second)
     alt_m = altitude.m_as(ureg.meter)
     a = _speed_of_sound_ms(alt_m)
-    return tas_ms / a  # type: ignore[no-any-return]
+    return float(tas_ms / a)
